Scheduler: make accidental overload less possible.
I believe I have seen on multiple occasions, an invocation such as:

    foo.Schedule([] (error *err) -> void {}, err);

This was matching against

    Schedule(std::function<>, bool sync=false, error *err=nullptr, ...)

With sync=err, interpreted as a bool ... So a synchronous enqueue!  Which
is pointless for most scenarios and often harmful, not what was intended,
etc.

Take some optionals out of the header for this to work as intended.