refcount.h: weak pointers
Inspired by std::weak_ptr.

If no weak pointers exist, the only overhead is an extra pointer in the RefCountable
objects.

The first caller to request a weak pointer must do memory allocation, allocating a
"weak pointer control block", which is refcounted, and contains an rwlock.

Thereon after, RefCountable::Release acquires the rwlock in shared mode, and nulls
out the weak back-pointer when the RefCountable is about to be destroyed.

Converting from weak to strong acquires the rwlock in exclusive mode, to ensure
the object won't be destroyed in RefCountable::Release before we bump up the
refcount.