Refactor "waiter node" concept.
I realized that both recmutex and rwlock-self had this blocking primitive.
On Unix, the former used semaphores and the latter used pipes or eventfd.

Factor out the latter implementation and point both at the same.

Maybe the semaphore implementation was better because most modern semaphore
implementations don't do a syscall if already signalled, and this one will
unconditionally enter read(2) to block.  So perhaps that is something to
re-visit.  However, I also like this not depending on semaphores, so that if
we need to re-implement semaphore in terms of this it's ok.